move to REST types to `api::rest::schema`, decouple `OrderByInterface` and `RecommendStrategy`
