+++ lib/api/src/grpc/conversions.rs
@@ -1,4 +1,5 @@
 use std::collections::{HashMap, HashSet};
+use std::str::FromStr as _;
 use std::time::Instant;
 
 use chrono::{NaiveDateTime, Timelike};
@@ -14,9 +15,7 @@ use uuid::Uuid;
 
 use super::qdrant::raw_query::RawContextPair;
 use super::qdrant::{
-    raw_query, start_from, BinaryQuantization, CompressionRatio, DatetimeRange, Direction,
-    GeoLineString, GroupId, MultiVectorComparator, MultiVectorConfig, OrderBy, OrderValue, Range,
-    RawVector, SparseIndices, StartFrom,
+    raw_query, start_from, BinaryQuantization, CompressionRatio, DatetimeRange, Direction, GeoLineString, GroupId, MultiVectorComparator, MultiVectorConfig, OrderBy, OrderValue, Range, RawVector, RecommendStrategy, SparseIndices, StartFrom
 };
 use crate::grpc::models::{CollectionsResponse, VersionInfo};
 use crate::grpc::qdrant::condition::ConditionOneOf;
@@ -1440,6 +1439,48 @@ impl From<segment::data_types::order_by::Direction> for Direction {
         }
     }
 }
+impl TryFrom<OrderBy> for segment::data_types::order_by::OrderBy {
+    type Error = Status;
+
+    fn try_from(value: OrderBy) -> Result<Self, Self::Error> {
+        use crate::grpc::qdrant::start_from::Value;
+        use segment::data_types::order_by::StartFrom;
+
+        let direction = value
+            .direction
+            .and_then(Direction::from_i32)
+            .map(segment::data_types::order_by::Direction::from);
+
+        let start_from = value
+            .start_from
+            .and_then(|value| value.value)
+            .map(|v| -> Result<StartFrom, Status> {
+                match v {
+                    Value::Integer(int) => {
+                        Ok(StartFrom::Integer(int))
+                    }
+                    Value::Float(float) => {
+                        Ok(StartFrom::Float(float))
+                    }
+                    Value::Timestamp(timestamp) => {
+                        Ok(StartFrom::Datetime(try_date_time_from_proto(timestamp)?))
+                    }
+                    Value::Datetime(datetime_str) => Ok(
+                        StartFrom::Datetime(segment::types::DateTimeWrapper::from_str(&datetime_str).map_err(
+                            |e| Status::invalid_argument(format!("Malformed datetime: {e}")),
+                        )?),
+                    ),
+                }
+            })
+            .transpose()?;
+
+        Ok(Self {
+            key: json_path_from_proto(&value.key)?,
+            direction,
+            start_from,
+        })
+    }
+}
 
 impl From<segment::data_types::order_by::OrderBy> for OrderBy {
     fn from(value: segment::data_types::order_by::OrderBy) -> Self {
@@ -1700,6 +1741,26 @@ impl TryFrom<RawVector> for segment_vectors::Vector {
 impl From<segment_vectors::NamedVectorStruct> for RawVector {
     fn from(value: segment_vectors::NamedVectorStruct) -> Self {
         Self::from(value.to_vector())
+    }
+}
+
+impl From<RecommendStrategy> for crate::rest::RecommendStrategy {
+    fn from(value: RecommendStrategy) -> Self {
+        match value {
+            RecommendStrategy::AverageVector => crate::rest::RecommendStrategy::AverageVector,
+            RecommendStrategy::BestScore => crate::rest::RecommendStrategy::BestScore,
+        }
+    }
+}
+
+impl TryFrom<i32> for crate::rest::RecommendStrategy {
+    type Error = Status;
+
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
+        let strategy = RecommendStrategy::from_i32(value).ok_or_else(|| {
+            Status::invalid_argument(format!("Unknown recommend strategy: {}", value))
+        })?;
+        Ok(strategy.into())
     }
 }
