+++ lib/api/src/rest/conversions.rs
@@ -1,4 +1,6 @@
-use super::schema::{BatchVectorStruct, ScoredPoint, Vector, VectorStruct};
+use segment::data_types::order_by::OrderBy;
+
+use super::{schema::{BatchVectorStruct, ScoredPoint, Vector, VectorStruct}, OrderByInterface};
 use crate::rest::{DenseVector, NamedVectorStruct};
 
 impl From<segment::data_types::vectors::Vector> for Vector {
@@ -164,5 +166,18 @@ impl From<DenseVector> for NamedVectorStruct {
 impl From<segment::data_types::vectors::NamedVector> for NamedVectorStruct {
     fn from(v: segment::data_types::vectors::NamedVector) -> Self {
         NamedVectorStruct::Dense(v)
+    }
+}
+
+impl From<OrderByInterface> for OrderBy {
+    fn from(order_by: OrderByInterface) -> Self {
+        match order_by {
+            OrderByInterface::Key(key) => OrderBy {
+                key,
+                direction: None,
+                start_from: None,
+            },
+            OrderByInterface::Struct(order_by) => order_by,
+        }
     }
 }
