+++ lib/collection/src/operations/universal_query/mod.rs
@@ -12,5 +12,4 @@
 
 pub mod collection_query;
 pub mod planned_query;
-pub mod rest;
 pub mod shard_query;
