+++ lib/storage/src/rbac/ops_checks.rs
@@ -525,16 +525,14 @@ mod tests {
 mod tests_ops {
     use std::fmt::Debug;
 
-    use api::rest::{BatchVectorStruct, VectorStruct};
+    use api::rest::{BatchVectorStruct, OrderByInterface, RecommendStrategy, VectorStruct};
     use collection::operations::payload_ops::PayloadOpsDiscriminants;
     use collection::operations::point_ops::{
         Batch, PointInsertOperationsInternal, PointInsertOperationsInternalDiscriminants,
         PointOperationsDiscriminants, PointStruct, PointSyncOperation,
     };
     use collection::operations::query_enum::QueryEnum;
-    use collection::operations::types::{
-        OrderByInterface, RecommendStrategy, SearchRequestInternal, UsingVector,
-    };
+    use collection::operations::types::{SearchRequestInternal, UsingVector};
     use collection::operations::vector_ops::{
         PointVectors, UpdateVectorsOp, VectorOperationsDiscriminants,
     };
